Addressing PR comments #1
