+++ tls/extensions/s2n_client_pq_kem.c
@@ -36,8 +36,7 @@ int s2n_extensions_client_pq_kem_send(struct s2n_connection *conn, struct s2n_st
     GUARD(s2n_connection_get_security_policy(conn, &security_policy));
     notnull_check(security_policy);
     /* Each supported kem id is 2 bytes */
-    const struct s2n_kem_preferences *kem_preferences = NULL;
-    kem_preferences = security_policy->kem_preferences;
+    const struct s2n_kem_preferences *kem_preferences = security_policy->kem_preferences;
     notnull_check(kem_preferences);
     for (int i = 0; i < kem_preferences->count; i++) {
         GUARD(s2n_stuffer_write_uint16(out, kem_preferences->kems[i]->kem_extension_id));
